Merging seaice_growth codes -- part 2 -- actual changes.
- change the units of d_HSNWbyOCNonSNW to snow meters.

In warming condition, when ice gets melted, a 1/area
factor appeared to be missing. It is added in V1.82.

This revision is not just a cosmetic one, since
it would change results in warming conditions.

Yet the test experiments results were not changed,
since they all start in January. It appears that
the block of code that does this ice melting is not tested!

I included a comment in seaice_growth.F in this regard. Ultimately
we should set some of the experiments to start e.g. in the spring.